:recycle: refactor : 패키지 구조를 application, domain, infra 영역으로 나누도록 수정
