+++ src/main/kotlin/io/hhplus/tdd/point/domain/UserPoint.kt
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point
+package io.hhplus.tdd.point.domain
 
 data class UserPoint(
     val id: Long,
