+++ src/main/kotlin/io/hhplus/tdd/point/infrastructure/web/PointController.kt
@@ -1,5 +1,7 @@
-package io.hhplus.tdd.point
+package io.hhplus.tdd.point.infrastructure.web
 
+import io.hhplus.tdd.point.domain.PointHistory
+import io.hhplus.tdd.point.domain.UserPoint
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.web.bind.annotation.*
@@ -15,9 +17,7 @@ class PointController {
     @GetMapping("{id}")
     fun point(
         @PathVariable id: Long,
-    ): UserPoint {
-        return UserPoint(0, 0, 0)
-    }
+    ): UserPoint = UserPoint(0, 0, 0)
 
     /**
      * TODO - 특정 유저의 포인트 충전/이용 내역을 조회하는 기능을 작성해주세요.
@@ -25,9 +25,7 @@ class PointController {
     @GetMapping("{id}/histories")
     fun history(
         @PathVariable id: Long,
-    ): List<PointHistory> {
-        return emptyList()
-    }
+    ): List<PointHistory> = emptyList()
 
     /**
      * TODO - 특정 유저의 포인트를 충전하는 기능을 작성해주세요.
@@ -36,9 +34,7 @@ class PointController {
     fun charge(
         @PathVariable id: Long,
         @RequestBody amount: Long,
-    ): UserPoint {
-        return UserPoint(0, 0, 0)
-    }
+    ): UserPoint = UserPoint(0, 0, 0)
 
     /**
      * TODO - 특정 유저의 포인트를 사용하는 기능을 작성해주세요.
@@ -47,7 +43,5 @@ class PointController {
     fun use(
         @PathVariable id: Long,
         @RequestBody amount: Long,
-    ): UserPoint {
-        return UserPoint(0, 0, 0)
-    }
+    ): UserPoint = UserPoint(0, 0, 0)
 }
