+++ src/main/kotlin/io/hhplus/tdd/point/infrastructure/web/config/ApiControllerAdvice.kt
@@ -1,4 +1,4 @@
-package io.hhplus.tdd
+package io.hhplus.tdd.point.infrastructure.web.config
 
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -8,17 +8,19 @@ import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.RestControllerAdvice
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler
 
-data class ErrorResponse(val code: String, val message: String)
+data class ErrorResponse(
+    val code: String,
+    val message: String,
+)
 
 @RestControllerAdvice
 class ApiControllerAdvice : ResponseEntityExceptionHandler() {
     private val logger: Logger = LoggerFactory.getLogger(javaClass)
 
     @ExceptionHandler(Exception::class)
-    fun handleException(e: Exception): ResponseEntity<ErrorResponse> {
-        return ResponseEntity(
+    fun handleException(e: Exception): ResponseEntity<ErrorResponse> =
+        ResponseEntity(
             ErrorResponse("500", "에러가 발생했습니다."),
             HttpStatus.INTERNAL_SERVER_ERROR,
         )
-    }
 }
